python_rag/requirements.txt
google-generativeai>=0.7.0
sentence-transformers>=2.5.0
psycopg2-binary>=2.9.9
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
pgvector>=0.2.0
jinja2>=3.1.0
tiktoken>=0.5.0
tree-sitter>=0.20.4
tree-sitter-languages>=1.10.0
